% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centroids.R
\name{centroids}
\alias{centroids}
\alias{centroids,ANY-method}
\alias{centroids,SpatialDataLabel-method}
\alias{centroids,SpatialDataShape-method}
\alias{centroids,SpatialDataPoint-method}
\title{Spatial element centroids}
\usage{
\S4method{centroids}{ANY}(x, ...)

\S4method{centroids}{SpatialDataLabel}(x, as = c("data.frame", "matrix"))

\S4method{centroids}{SpatialDataShape}(x, as = c("data.frame", "matrix", "list"))

\S4method{centroids}{SpatialDataPoint}(x, as = c("data.frame", "list"))
}
\arguments{
\item{x}{a \code{SpatialData} element (any but image).}

\item{...}{ignored.}

\item{as}{character string; how results should be returned.}
}
\value{
A table (\code{data.frame} or \code{matrix}) of spatial coordinates 
(if \code{as="list"}, split by instance (shapes) or features (points)).
}
\description{
Spatial element centroids
}
\examples{
x <- file.path("extdata", "blobs.zarr")
x <- system.file(x, package="spatialdataR")
x <- readSpatialData(x, tables=FALSE)

centroids(label(x))
centroids(shape(x))

head(centroids(point(x)))
xy <- centroids(point(x), "list")
plot(xy$gene_a, col=a <- "red")
points(xy$gene_b, col=b <- "blue")
legend("topright", legend=names(xy), col=c(a, b), pch=21)
}
