% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/methods.R
\docType{class}
\name{CoordinateTransform-class}
\alias{CoordinateTransform-class}
\alias{.CoordinateTransform}
\alias{show,CoordinateTransform-method}
\title{CoordinateTransform: Spatial coordinate transformation}
\usage{
\S4method{show}{CoordinateTransform}(object)
}
\arguments{
\item{object}{A \code{CoordinateTransform} object.}
}
\description{
Represents an affine transformation between coordinate systems.
}
\section{Slots}{

\describe{
\item{\code{type}}{Character. \code{"identity"} or \code{"affine"}.}

\item{\code{affine}}{Numeric matrix (3x3 for 2D).}

\item{\code{input_cs}}{Character. Input coordinate system name.}

\item{\code{output_cs}}{Character. Output coordinate system name.}
}}

\references{
Marconato L et al. (2024). SpatialData: an open and universal
data framework for spatial omics. \emph{Nat Methods} 21:2196-2209.
\doi{10.1038/s41592-024-02212-x}
}
