% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{.addImageLayer}
\alias{.addImageLayer}
\title{Add image background layer}
\usage{
.addImageLayer(p, sdata, image_name)
}
\arguments{
\item{p}{ggplot object.}

\item{sdata}{SpatialData object.}

\item{image_name}{Character. Name of image element.}
}
\value{
Updated ggplot object.
}
\description{
Add image background layer
}
\keyword{internal}
