% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{.addPointsLayer}
\alias{.addPointsLayer}
\title{Add points layer}
\usage{
.addPointsLayer(
  p,
  sdata,
  points_name,
  color_by = NULL,
  point_size = 0.3,
  point_alpha = 0.5,
  max_points = 500000L
)
}
\arguments{
\item{p}{ggplot object.}

\item{sdata}{SpatialData object.}

\item{points_name}{Character. Name of points element.}

\item{color_by}{Column name for coloring.}

\item{point_size}{Point size.}

\item{point_alpha}{Point alpha.}

\item{max_points}{Maximum points to plot.}
}
\value{
Updated ggplot object.
}
\description{
Add points layer
}
\keyword{internal}
