% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{.addShapesLayer}
\alias{.addShapesLayer}
\title{Add shapes layer}
\usage{
.addShapesLayer(p, sdata, shapes_name, fill = "transparent", color = "red")
}
\arguments{
\item{p}{ggplot object.}

\item{sdata}{SpatialData object.}

\item{shapes_name}{Character. Name of shapes element.}

\item{fill}{Fill color for polygons.}

\item{color}{Border color for polygons.}
}
\value{
Updated ggplot object.
}
\description{
Add shapes layer
}
\keyword{internal}
