% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{.buildCountMatrix}
\alias{.buildCountMatrix}
\title{Build feature-by-region count matrix}
\usage{
.buildCountMatrix(features, region_ids, valid_regions)
}
\arguments{
\item{features}{Character vector of feature IDs.}

\item{region_ids}{Vector of region assignments.}

\item{valid_regions}{Vector of valid region IDs.}
}
\value{
Integer matrix.
}
\description{
Build feature-by-region count matrix
}
\keyword{internal}
