% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial-stats.R
\name{.buildKNN}
\alias{.buildKNN}
\title{Build k-nearest-neighbor index}
\usage{
.buildKNN(coords, k)
}
\arguments{
\item{coords}{data.frame with x, y.}

\item{k}{Number of neighbors.}
}
\value{
List with nn.index and nn.dist matrices.
}
\description{
Build k-nearest-neighbor index
}
\keyword{internal}
