% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{.checkDataPresence}
\alias{.checkDataPresence}
\title{Check data file presence for an element}
\usage{
.checkDataPresence(elem_path, etype, ename)
}
\arguments{
\item{elem_path}{Element directory path.}

\item{etype}{Element type.}

\item{ename}{Element name.}
}
\value{
List with has_data and warnings.
}
\description{
Check data file presence for an element
}
\keyword{internal}
