% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-elements.R
\name{.discoverZarrColumns}
\alias{.discoverZarrColumns}
\title{Discover column order from Zarr group .zattrs}
\usage{
.discoverZarrColumns(group_path)
}
\arguments{
\item{group_path}{Path to a Zarr group.}
}
\value{
Character vector of column names.
}
\description{
Discover column order from Zarr group .zattrs
}
\keyword{internal}
