% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce-bioc.R
\name{.dropIndexCol}
\alias{.dropIndexCol}
\title{Remove index columns from a DataFrame for colData/rowData}
\usage{
.dropIndexCol(df)
}
\arguments{
\item{df}{A DataFrame.}
}
\value{
DataFrame with index columns removed.
}
\description{
Remove index columns from a DataFrame for colData/rowData
}
\keyword{internal}
