% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce-bioc.R
\name{.extractIndex}
\alias{.extractIndex}
\title{Extract index column values from a DataFrame}
\usage{
.extractIndex(df)
}
\arguments{
\item{df}{A DataFrame that may have an X_index or _index column.}
}
\value{
Character vector or NULL.
}
\description{
Extract index column values from a DataFrame
}
\keyword{internal}
