% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce-bioc.R
\name{.getCoordsFromShapes}
\alias{.getCoordsFromShapes}
\title{Get spatial coordinates from shape centroids}
\usage{
.getCoordsFromShapes(sdata, shapes_name, n_cells)
}
\arguments{
\item{sdata}{SpatialData object.}

\item{shapes_name}{Name of shapes element (or NULL for auto).}

\item{n_cells}{Expected number of cells.}
}
\value{
A matrix with x, y columns or NULL.
}
\description{
Get spatial coordinates from shape centroids
}
\keyword{internal}
