% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-zarr.R
\name{.loadTabularElement}
\alias{.loadTabularElement}
\title{Load tabular data from an element directory}
\usage{
.loadTabularElement(elem_path)
}
\arguments{
\item{elem_path}{Path to element directory.}
}
\value{
A \code{DataFrame} or \code{NULL}.
}
\description{
Load tabular data from an element directory
}
\keyword{internal}
