% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multisample.R
\name{.mergeCoordSystems}
\alias{.mergeCoordSystems}
\title{Merge coordinate systems with sample prefixes}
\usage{
.mergeCoordSystems(objects, sample_ids, n)
}
\arguments{
\item{objects}{List of SpatialData.}

\item{sample_ids}{Character vector.}

\item{n}{Number of objects.}
}
\value{
Named list.
}
\description{
Merge coordinate systems with sample prefixes
}
\keyword{internal}
