% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial-stats.R
\name{.moranVariance}
\alias{.moranVariance}
\title{Compute asymptotic standard deviation of Moran's I}
\usage{
.moranVariance(n, k, S0)
}
\arguments{
\item{n}{Number of observations.}

\item{k}{Number of neighbors.}

\item{S0}{Sum of weights.}
}
\value{
Numeric standard deviation.
}
\description{
Compute asymptotic standard deviation of Moran's I
}
\keyword{internal}
