% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{.normalizeCounts}
\alias{.normalizeCounts}
\title{Normalize count matrix by region sizes}
\usage{
.normalizeCounts(mat, region_ids)
}
\arguments{
\item{mat}{Count matrix.}

\item{region_ids}{Region assignments.}
}
\value{
Normalized matrix.
}
\description{
Normalize count matrix by region sizes
}
\keyword{internal}
