% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{.parseOneWKB}
\alias{.parseOneWKB}
\title{Parse a single WKB geometry}
\usage{
.parseOneWKB(wkb)
}
\arguments{
\item{wkb}{Raw vector.}
}
\value{
Parsed geometry (data.frame for Point, list for
  Polygon/MultiPolygon).
}
\description{
Parse a single WKB geometry
}
\keyword{internal}
