% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{.parseTransform}
\alias{.parseTransform}
\title{Parse transformation from Zarr metadata}
\usage{
.parseTransform(metadata)
}
\arguments{
\item{metadata}{List. Parsed \code{.zattrs} content.}
}
\value{
A \code{\link{CoordinateTransform}} or \code{NULL}.
}
\description{
Reads coordinate transformation definitions from
SpatialData element \code{.zattrs} metadata.
}
\keyword{internal}
