% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{.parseWKBMultiPolygon}
\alias{.parseWKBMultiPolygon}
\title{Parse a single WKB MultiPolygon}
\usage{
.parseWKBMultiPolygon(raw_bytes, offset, endian = "little")
}
\arguments{
\item{raw_bytes}{Raw vector.}

\item{offset}{1-based start (after type).}

\item{endian}{Endianness.}
}
\value{
List of lists of coordinate matrices.
}
\description{
Parse a single WKB MultiPolygon
}
\keyword{internal}
