% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{.parseWKBPoint}
\alias{.parseWKBPoint}
\title{Parse a single WKB Point}
\usage{
.parseWKBPoint(raw_bytes, offset, endian = "little")
}
\arguments{
\item{raw_bytes}{Raw vector.}

\item{offset}{1-based start of the point data (after type).}

\item{endian}{Endianness.}
}
\value{
data.frame with x, y.
}
\description{
Parse a single WKB Point
}
\keyword{internal}
