% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{.parseWKBPolygon}
\alias{.parseWKBPolygon}
\title{Parse a single WKB Polygon}
\usage{
.parseWKBPolygon(raw_bytes, offset, endian = "little")
}
\arguments{
\item{raw_bytes}{Raw vector.}

\item{offset}{1-based start of polygon data (after type).}

\item{endian}{Endianness.}
}
\value{
List of coordinate matrices (one per ring).
}
\description{
Parse a single WKB Polygon
}
\keyword{internal}
