% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial-join.R
\name{.polyBBox}
\alias{.polyBBox}
\title{Compute bounding box of polygon vertices}
\usage{
.polyBBox(poly_x, poly_y)
}
\arguments{
\item{poly_x}{Numeric vector of x coordinates.}

\item{poly_y}{Numeric vector of y coordinates.}
}
\value{
Named numeric: xmin, xmax, ymin, ymax.
}
\description{
Compute bounding box of polygon vertices
}
\keyword{internal}
