% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial-query.R
\name{.querySimpleList}
\alias{.querySimpleList}
\title{Query a SimpleList of mixed elements}
\usage{
.querySimpleList(sl, xmin, xmax, ymin, ymax)
}
\arguments{
\item{sl}{SimpleList of DataFrames or path refs.}

\item{xmin, xmax, ymin, ymax}{Bounding box.}
}
\value{
Queried SimpleList.
}
\description{
Query a SimpleList of mixed elements
}
\keyword{internal}
