% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-elements.R
\name{.readAnnDataGroup}
\alias{.readAnnDataGroup}
\title{Read an AnnData obs/var group (Zarr arrays or CSV)}
\usage{
.readAnnDataGroup(group_path)
}
\arguments{
\item{group_path}{Path to the obs or var Zarr group.}
}
\value{
A \code{DataFrame}.
}
\description{
Read an AnnData obs/var group (Zarr arrays or CSV)
}
\keyword{internal}
