% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{.readFloat64}
\alias{.readFloat64}
\title{Read float64 (double) from raw bytes}
\usage{
.readFloat64(raw_bytes, offset, endian = "little")
}
\arguments{
\item{raw_bytes}{Raw vector.}

\item{offset}{1-based position to start reading.}

\item{endian}{"little" or "big".}
}
\value{
Numeric value.
}
\description{
Read float64 (double) from raw bytes
}
\keyword{internal}
