% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-elements.R
\name{.readInt64Raw}
\alias{.readInt64Raw}
\title{Get R type info for a Zarr v3 data_type string}
\usage{
.readInt64Raw(raw_data, n, endian = "little")
}
\arguments{
\item{dtype}{Character. Zarr v3 data type string.}
}
\value{
A list with \code{what}, \code{size}, \code{signed}.
}
\description{
Get R type info for a Zarr v3 data_type string
}
\keyword{internal}
