% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{.readUint32}
\alias{.readUint32}
\title{Read uint32 from raw bytes}
\usage{
.readUint32(raw_bytes, offset, endian = "little")
}
\arguments{
\item{raw_bytes}{Raw vector.}

\item{offset}{1-based position to start reading.}

\item{endian}{"little" or "big".}
}
\value{
Integer value.
}
\description{
Read uint32 from raw bytes
}
\keyword{internal}
