% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-elements.R
\name{.readZarrColumns}
\alias{.readZarrColumns}
\title{Read Zarr columnar arrays into a DataFrame}
\usage{
.readZarrColumns(group_path, col_order)
}
\arguments{
\item{group_path}{Path to a Zarr group.}

\item{col_order}{Character vector of column names.}
}
\value{
A \code{DataFrame}.
}
\description{
Read Zarr columnar arrays into a DataFrame
}
\keyword{internal}
