% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{.validateAllElements}
\alias{.validateAllElements}
\title{Validate all elements in store}
\usage{
.validateAllElements(path)
}
\arguments{
\item{path}{Zarr store root.}
}
\value{
List with errors, warnings, elements df.
}
\description{
Validate all elements in store
}
\keyword{internal}
