% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{.validateTopZattrs}
\alias{.validateTopZattrs}
\title{Validate top-level .zattrs}
\usage{
.validateTopZattrs(path)
}
\arguments{
\item{path}{Zarr store root.}
}
\value{
List with errors and warnings.
}
\description{
Validate top-level .zattrs
}
\keyword{internal}
