% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{.wkbEndian}
\alias{.wkbEndian}
\title{Determine endianness from WKB byte}
\usage{
.wkbEndian(byte)
}
\arguments{
\item{byte}{Single raw byte.}
}
\value{
Character: "little" or "big".
}
\description{
Determine endianness from WKB byte
}
\keyword{internal}
