% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{.wkbTypeName}
\alias{.wkbTypeName}
\title{WKB type ID to type name}
\usage{
.wkbTypeName(type_id)
}
\arguments{
\item{type_id}{Integer WKB type code.}
}
\value{
Character type name.
}
\description{
WKB type ID to type name
}
\keyword{internal}
