% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-zarr.R
\name{.writeDataFrameElements}
\alias{.writeDataFrameElements}
\title{Write DataFrame elements (points/shapes)}
\usage{
.writeDataFrameElements(sl, base_path, type_dir)
}
\arguments{
\item{sl}{SimpleList of DataFrames or refs.}

\item{base_path}{Zarr store root.}

\item{type_dir}{"points" or "shapes".}
}
\value{
Invisible \code{NULL}.
}
\description{
Write DataFrame elements (points/shapes)
}
\keyword{internal}
