% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-zarr.R
\name{.writeRefElements}
\alias{.writeRefElements}
\title{Write image/label references}
\usage{
.writeRefElements(sl, base_path, type_dir)
}
\arguments{
\item{sl}{SimpleList of element refs.}

\item{base_path}{Zarr store root.}

\item{type_dir}{"images" or "labels".}
}
\description{
Write image/label references
}
\keyword{internal}
