% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write-zarr.R
\name{.writeTableElements}
\alias{.writeTableElements}
\title{Write table elements}
\usage{
.writeTableElements(sl, base_path)
}
\arguments{
\item{sl}{SimpleList of table objects.}

\item{base_path}{Zarr store root.}
}
\value{
Invisible \code{NULL}.
}
\description{
Write table elements
}
\keyword{internal}
