% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-elements.R
\name{.zarrV3TypeInfo}
\alias{.zarrV3TypeInfo}
\title{Get R type info for a Zarr v3 data_type string}
\usage{
.zarrV3TypeInfo(dtype)
}
\arguments{
\item{dtype}{Character. Zarr v3 data type string.}
}
\value{
A list with \code{what}, \code{size}, \code{signed}.
}
\description{
Get R type info for a Zarr v3 data_type string
}
\keyword{internal}
