% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delayed.R
\name{loadElement}
\alias{loadElement}
\title{Load element as DelayedArray from SpatialData}
\usage{
loadElement(sd, element_name, scale = "scale0")
}
\arguments{
\item{sd}{A \code{\linkS4class{SpatialData}} object.}

\item{element_name}{Character. Name of the image or
label element.}

\item{scale}{Character. Multiscale level to load
(default: \code{"scale0"}).}
}
\value{
A \code{DelayedArray} or regular array.
}
\description{
Convenience wrapper to load an image or label
element from a \code{SpatialData} object as a
\code{DelayedArray}.
}
\examples{
store <- system.file("extdata", "xenium_mini.zarr",
    package = "SpatialDataR")
sd <- readSpatialData(store)
# Load morphology image (if Rarr available):
# img <- loadElement(sd, "morphology")
}
