% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{path}
\alias{path}
\alias{path,SpatialDataArray-method}
\alias{path,SpatialDataFrame-method}
\alias{path,SingleCellExperiment-method}
\alias{path,SpatialData-method}
\title{Retrieve \code{SpatialData} on-disk paths}
\usage{
\S4method{path}{SpatialDataArray}(object, ...)

\S4method{path}{SpatialDataFrame}(object, ...)

\S4method{path}{SingleCellExperiment}(object, ...)

\S4method{path}{SpatialData}(object, simplify = TRUE, ...)
}
\arguments{
\item{object}{\code{\link{SpatialData}} object or one of its elements.}

\item{...}{ignored.}

\item{simplify}{logical scalar; whether to flatten paths into a tibble.}
}
\value{
for single elements, a character string; 
for \link{SpatialData} objects, if \code{simplify=TRUE} (default), 
a \code{tibble} where rows=elements and columns=layers/elements/paths. 
if \code{simplify=FALSE}, a depth-3 list where levels=layers/elements/paths.
}
\description{
Retrieve \code{SpatialData} on-disk paths
}
\examples{
zs <- file.path("extdata", "blobs.zarr")
zs <- system.file(zs, package="spatialdataR")
sd <- readSpatialData(zs)

# element-wise
path(shape(sd))

# object-wide
path(sd)
path(sd, FALSE)$labels

}
