% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{readSpatialData}
\alias{readSpatialData}
\alias{readImage}
\alias{readLabel}
\alias{readPoint}
\alias{readShape}
\alias{readTable}
\title{Reading `SpatialData`}
\usage{
readImage(x, ...)

readLabel(x, ...)

readPoint(x, ...)

readShape(x, ...)

readTable(x)

readSpatialData(
  x,
  images = TRUE,
  labels = TRUE,
  points = TRUE,
  shapes = TRUE,
  tables = TRUE
)
}
\arguments{
\item{x}{For \code{readImage/Label/Point/Shape/Table},
path to a \code{SpatialData} element.
For \code{readSpatialData},
path to a \code{SpatialData}-.zarr store.}

\item{...}{option arguments passed to and from other methods.}

\item{images, labels, points, shapes, tables}{Control which elements should be read for each layer.
The default, NULL, reads all elements; alternatively, may be FALSE
to skip a layer, or a integer vector specifying which elements to read.}
}
\value{
\itemize{
\item{For \code{readSpatialData}, a \code{SpatialData}.},
\item{For element readers, 
a \code{SpatialDataImage/Label/Point/Shape} 
or \code{SingleCellExperiment}.}}
}
\description{
Reading `SpatialData`
}
\examples{
zs <- file.path("extdata", "blobs.zarr")
zs <- system.file(zs, package="spatialdataR")

# read complete Zarr store
(sd <- readSpatialData(zs))

# helper that gets path to last element in layer 'l'
fn <- \(.) tail(list.files(file.path(zs, .), full.names=TRUE), 1)

# read individual elements
(i <- readImage(fn("images")))
channels(i)

(p <- readPoint(fn("points")))
as.data.frame(head(p))

(s <- readShape(fn("shapes")))
data(s)
}
