% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods.R
\name{images}
\alias{images}
\alias{spatialLabels}
\alias{spatialPoints}
\alias{shapes}
\alias{tables}
\alias{coordinateSystems}
\alias{images,SpatialData-method}
\alias{spatialLabels,SpatialData-method}
\alias{spatialPoints,SpatialData-method}
\alias{shapes,SpatialData-method}
\alias{tables,SpatialData-method}
\alias{coordinateSystems,SpatialData-method}
\title{Access images from a SpatialData object}
\usage{
images(x)

spatialLabels(x)

spatialPoints(x)

shapes(x)

tables(x)

coordinateSystems(x)

\S4method{images}{SpatialData}(x)

\S4method{spatialLabels}{SpatialData}(x)

\S4method{spatialPoints}{SpatialData}(x)

\S4method{shapes}{SpatialData}(x)

\S4method{tables}{SpatialData}(x)

\S4method{coordinateSystems}{SpatialData}(x)
}
\arguments{
\item{x}{A \code{\linkS4class{SpatialData}} object.}
}
\value{
A \code{SimpleList} of image references.

A \code{SimpleList} of label references.

A \code{SimpleList} of point DataFrames.

A \code{SimpleList} of shape DataFrames.

A \code{SimpleList} of \code{SpatialExperiment} objects.

A named list of coordinate system definitions.
}
\description{
Access images from a SpatialData object

Access labels/segmentation masks

Access point coordinates

Access shape geometries

Access annotation tables

Access coordinate systems
}
\examples{
sd <- new("SpatialData")
images(sd)
spatialLabels(sd)
spatialPoints(sd)
shapes(sd)
tables(sd)
}
