% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/methods.R
\docType{class}
\name{SpatialData-class}
\alias{SpatialData-class}
\alias{.SpatialData}
\alias{show,SpatialData-method}
\title{SpatialData: Container for multi-modal spatial omics data}
\usage{
\S4method{show}{SpatialData}(object)
}
\arguments{
\item{object}{A \code{SpatialData} object.}
}
\description{
A lightweight R-native representation of a SpatialData Zarr
store. Holds references to images, labels, points, shapes,
and annotation tables without loading full data into memory.
}
\section{Slots}{

\describe{
\item{\code{images}}{\code{SimpleList} of image references.}

\item{\code{labels}}{\code{SimpleList} of label/segmentation references.}

\item{\code{points}}{\code{SimpleList} of point DataFrames.}

\item{\code{shapes}}{\code{SimpleList} of shape DataFrames.}

\item{\code{tables}}{\code{SimpleList} of annotation tables.}

\item{\code{coordinate_systems}}{\code{list} of coordinate systems.}

\item{\code{metadata}}{\code{list} of additional metadata.}

\item{\code{path}}{Character. Path to the source Zarr store.}
}}

\examples{
showClass("SpatialData")
}
\references{
Marconato L et al. (2024). SpatialData: an open and universal
data framework for spatial omics. \emph{Nat Methods} 21:2196-2209.
\doi{10.1038/s41592-024-02212-x}
}
