% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdArray.R
\name{SpatialDataArray}
\alias{SpatialDataArray}
\alias{data_type}
\alias{data,SpatialDataArray-method}
\alias{dim,SpatialDataArray-method}
\alias{length,SpatialDataArray-method}
\alias{data_type,SpatialDataArray-method}
\alias{data_type,DelayedArray-method}
\title{\code{SpatialDataArray}}
\usage{
\S4method{data}{SpatialDataArray}(x, k = 1)

\S4method{dim}{SpatialDataArray}(x)

\S4method{length}{SpatialDataArray}(x)

\S4method{data_type}{SpatialDataArray}(x)

\S4method{data_type}{DelayedArray}(x)
}
\arguments{
\item{x}{\code{SpatialDataImage} or \code{SpatialDataLabel}.}

\item{k}{scalar index specifying which scale to extract.}
}
\value{
\code{SpatialDataArray}
}
\description{
\code{SpatialDataArray}
}
\examples{
zs <- file.path("extdata", "blobs.zarr")
zs <- system.file(zs, package="SpatialData")

pa <- list.dirs(
  file.path(zs, "images"), 
  recursive=FALSE, full.names=TRUE)

(x <- readImage(pa[2]))  

data_type(x)
dim(data(x, 1))   # highest res.
dim(data(x, Inf)) # lowest res.

rgb <- apply(
  data(x, 1), c(2, 3), 
  \(.) rgb(.[1], .[2], .[3]))
plot(
  row(rgb), col(rgb), col=rgb, 
  pch=15, asp=1, ylim=c(ncol(rgb), 0))

}
