% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdImage.R
\name{SpatialDataImage}
\alias{SpatialDataImage}
\alias{channels}
\alias{channels,SpatialDataAttrs-method}
\alias{channels,SpatialDataImage-method}
\alias{channels,SpatialDataElement-method}
\alias{[,SpatialDataImage,ANY,ANY,ANY-method}
\title{The `SpatialDataImage` class}
\usage{
SpatialDataImage(
  data = list(),
  meta = SpatialDataAttrs(),
  metadata = list(),
  ...
)

\S4method{channels}{SpatialDataAttrs}(x, ...)

\S4method{channels}{SpatialDataImage}(x, ...)

\S4method{channels}{SpatialDataElement}(x, ...)

\S4method{[}{SpatialDataImage,ANY,ANY,ANY}(x, i, j, k, ..., drop = FALSE)
}
\arguments{
\item{data}{list of \code{\link{SpatialDataAttrs}}s}

\item{meta}{\code{\link{SpatialDataAttrs}}}

\item{metadata}{optional list of arbitrary 
content describing the overall object.}

\item{...}{option arguments passed to and from other methods.}

\item{x}{\code{SpatialDataImage}}

\item{i, j}{indices specifying elements to extract.}

\item{k}{scalar index specifying which scale to extract.}

\item{drop}{ignored.}
}
\value{
\code{SpatialDataImage}
}
\description{
The `SpatialDataImage` class
}
\examples{
zs <- file.path("extdata", "blobs.zarr")
zs <- system.file(zs, package="SpatialData")

pa <- list.dirs(
  file.path(zs, "images"), 
  recursive=FALSE, full.names=TRUE)

# simple
readImage(pa[1])

# multi-scale
(x <- readImage(pa[2]))

channels(x)
dim(data(x, 1))   # highest res.
dim(data(x, Inf)) # lowest res.

rgb <- apply(
  data(x, 1), c(2, 3), 
  \(.) rgb(.[1], .[2], .[3]))
plot(
  row(rgb), col(rgb), col=rgb, 
  pch=15, asp=1, ylim=c(ncol(rgb), 0))
}
