% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdLabel.R
\name{SpatialDataLabel}
\alias{SpatialDataLabel}
\alias{[,SpatialDataLabel,ANY,ANY,ANY-method}
\title{\code{SpatialDataLabel}}
\usage{
SpatialDataLabel(
  data = list(),
  meta = SpatialDataAttrs(),
  metadata = list(),
  ...
)

\S4method{[}{SpatialDataLabel,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{data}{list of \code{\link[ZarrArray]{ZarrArray}}s}

\item{meta}{\code{\link{SpatialDataAttrs}}}

\item{metadata}{optional list of arbitrary 
content describing the overall object.}

\item{...}{option arguments passed to and from other methods.}

\item{x}{\code{SpatialDataLabel}}

\item{i, j}{indices specifying elements to extract.}

\item{drop}{ignored.}
}
\value{
\code{SpatialDataLabel}
}
\description{
The \code{SpatialDataLabel} class stores \code{SpatialData} elements from its 
\code{"labels"} layers. These are represented as a \code{ZarrMatrix} 
(\code{data} slot) associated with \code{\link{SpatialDataAttrs}} 
(\code{meta} slot); a list of \code{metadata} stores other arbitrary info.

Currently defined methods (here, \code{x} is a \code{SpatialDataLabel}):
\itemize{
\item \code{data/meta(x)} to access underlying \code{ZarrMatrix/SpatialDataAttrs}
\item \code{dim(x)} returns the dimensions of \code{data(x)}
}
}
\examples{
x <- file.path("extdata", "blobs.zarr")
x <- system.file(x, package="SpatialData")
x <- file.path(x, "labels", "blobs_labels")

(y <- readLabel(x))
y[1:10, 1:10]
meta(y)

}
