% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans.R
\name{trans}
\alias{trans}
\alias{transform}
\alias{scale}
\alias{rotate}
\alias{translation}
\alias{flip}
\alias{flop}
\alias{mirror}
\alias{sequence}
\alias{transform,SpatialDataElement-method}
\alias{sequence,SpatialDataElement,list-method}
\alias{mirror,SpatialDataArray-method}
\alias{flip,SpatialDataArray-method}
\alias{flop,SpatialDataArray-method}
\alias{rotate,SpatialDataArray,numeric-method}
\alias{scale,SpatialDataArray,numeric-method}
\alias{translation,SpatialDataArray,numeric-method}
\alias{rotate,SpatialDataFrame,numeric-method}
\alias{scale,SpatialDataFrame,numeric-method}
\alias{translation,SpatialDataFrame,numeric-method}
\title{Transformations}
\usage{
\S4method{transform}{SpatialDataElement}(x, i = 1, ...)

\S4method{sequence}{SpatialDataElement,list}(x, t, ..., rev = FALSE)

\S4method{mirror}{SpatialDataArray}(x, t = c("v", "h"), k = 1, ...)

\S4method{flip}{SpatialDataArray}(x, k = 1, ...)

\S4method{flop}{SpatialDataArray}(x, k = 1, ...)

\S4method{rotate}{SpatialDataArray,numeric}(x, t, k = 1, ..., rev = FALSE)

\S4method{scale}{SpatialDataArray,numeric}(x, t, ...)

\S4method{translation}{SpatialDataArray,numeric}(x, t, ...)

\S4method{rotate}{SpatialDataFrame,numeric}(x, t, ...)

\S4method{scale}{SpatialDataFrame,numeric}(x, t, ...)

\S4method{translation}{SpatialDataFrame,numeric}(x, t, ...)
}
\arguments{
\item{x}{\code{SpatialData} element.}

\item{i}{scalar integer or string; target coordinate space.}

\item{...}{option arguments passed to and from other methods.}

\item{t}{transformation data; exceptions: for \code{mirror}, controls
whether to perform \bold{v}ertical or \bold{h}orizontal reflection;
no data is needed for \code{flip} (\bold{v}) and \code{flop} (\bold{h}).}

\item{rev}{flag; should transformation(s) be reversed?}

\item{k}{scalar index specifying which scale to use; 
\code{Inf} to use lowest available resolution;
only applies to \code{SpatialDataArray}s (images, labels).}
}
\value{
\code{SpatialData} element with transformation(s) applied.
}
\description{
Transformations
}
\examples{
x <- file.path("extdata", "blobs.zarr")
x <- system.file(x, package="spatialdataR")
x <- readSpatialData(x, tables=FALSE)

# image
y <- x
image(y) <- scale(image(y), c(1, 1, 1/3))
dim(image(x))
dim(image(y))
  
# point
y <- x
point(y, "rot") <- rotate(point(y), 20)
point(y, "wide") <- scale(point(y), c(1.2, 1))

xy0 <- centroids(point(y))
xy1 <- centroids(point(y, "rot"))
xy2 <- centroids(point(y, "wide"))

plot(xy0[, c(1, 2)], asp=1)
points(xy1[, c(1, 2)], col=2)
points(xy2[, c(1, 2)], col=4)
  
# shape
y <- x
shape(y, "rot") <- rotate(shape(y), 5)
shape(y, "wide") <- scale(shape(y), c(1.2, 1))
shape(y, "left") <- translation(shape(y), c(-5, 0))
y["shapes", c("rot", "wide", "left")]
}
