% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTutils.R
\name{CTutils}
\alias{CTutils}
\alias{axes}
\alias{CTlist}
\alias{CTname}
\alias{CTtype}
\alias{CTdata}
\alias{addCT}
\alias{rmvCT}
\alias{axes,SpatialDataAttrs-method}
\alias{CTlist,SpatialDataAttrs-method}
\alias{CTdata,SpatialDataAttrs-method}
\alias{CTtype,SpatialDataAttrs-method}
\alias{CTname,SpatialDataAttrs-method}
\alias{axes,SpatialDataElement-method}
\alias{CTlist,SpatialDataElement-method}
\alias{CTtype,SpatialDataElement-method}
\alias{CTname,SpatialDataElement-method}
\alias{CTdata,SpatialDataElement-method}
\alias{CTname,SpatialData-method}
\alias{rmvCT,SpatialDataElement-method}
\alias{rmvCT,SpatialDataAttrs-method}
\alias{addCT,SpatialDataElement-method}
\alias{addCT,SpatialDataAttrs-method}
\title{Coord. trans. utilities}
\usage{
\S4method{axes}{SpatialDataAttrs}(x, y = NULL, ...)

\S4method{CTlist}{SpatialDataAttrs}(x, ...)

\S4method{CTdata}{SpatialDataAttrs}(x, i = 1, ...)

\S4method{CTtype}{SpatialDataAttrs}(x, ...)

\S4method{CTname}{SpatialDataAttrs}(x, ...)

\S4method{axes}{SpatialDataElement}(x, y = NULL, ...)

\S4method{CTlist}{SpatialDataElement}(x, ...)

\S4method{CTtype}{SpatialDataElement}(x, ...)

\S4method{CTname}{SpatialDataElement}(x, ...)

\S4method{CTdata}{SpatialDataElement}(x, i = 1, ...)

\S4method{CTname}{SpatialData}(x, ...)

\S4method{rmvCT}{SpatialDataElement}(x, i)

\S4method{rmvCT}{SpatialDataAttrs}(x, i)

\S4method{addCT}{SpatialDataElement}(x, name, type = "identity", data = NULL)

\S4method{addCT}{SpatialDataAttrs}(x, name, type = "identity", data = NULL)
}
\arguments{
\item{x}{\code{SpatialData}, an element, or \code{SpatialDataAttrs}.}

\item{y}{NULL (default) returns a list where each element is 
an axis: a list with name/type/unit (e.g., x/space/micrometer);
\code{y="name/type/unit"} extracts specific data over all axiis.}

\item{...}{option arguments passed to and from other methods.}

\item{i}{for \code{CTpath}, source node label; else, string or 
scalar integer giving the name or index of a coordinate space.}

\item{name}{character(1); name of coordinate space}

\item{type}{character(1); type of transformation}

\item{data}{transformation data; size and shape depend on transformation and
element type (e.g., numeric(1) for rotation, numeric(2) for scaling in 2D)}
}
\value{
\itemize{
\item \code{CTname}: character string; 
  transformation name (e.g., "global")
\item \code{CTtype}: character string; 
  transformation type (e.g., "affine")
\item \code{CTdata}: list;
  transformation data (e.g., scalar numeric for rotation)
\item \code{CTlist}: list;
  list of transformation specifications per OME-NGFF spec
\item \code{add/rmvCT}: 
  \code{SpatialDataElement} or \code{SpatialDataAttrs} 
  with transformation(s) added/removed
\item \code{axes}: list; 
  each element is a character string (name), or list 
  with axis name and type (e.g., "space" or "channel")
}
}
\description{
Coord. trans. utilities
}
\examples{
x <- file.path("extdata", "blobs.zarr")
x <- system.file(x, package="spatialdataR")
x <- readSpatialData(x, tables=FALSE)

# view available target coordinate systems
CTname(z <- meta(label(x)))

# add
addCT(z, "scale", "scale", c(12, 34)) # overwrite
CTname(addCT(z, "new", "translation", c(12, 34)))

# rmv
CTname(rmvCT(z, 2))        # by index
CTname(rmvCT(z, "scale"))  # by name
CTname(rmvCT(z, "global")) # identity is protected
}
